Add WP_Error factory for ease of unit testing
Separate functions into another file

Rename Validate_Merge_Field class file
