+++ includes/validation/class-mailchimp-validation.php
@@ -31,6 +31,7 @@ class Mailchimp_Validation {
 	 * @return void
 	 */
 	private function require_validation_functions() {
-		include_once MCSF_DIR . 'includes/validation/merge-fields.php';
+		include_once MCSF_DIR . 'includes/validation/class-validate-merge-fields.php';
+		include_once MCSF_DIR . 'includes/validation/functions.php';
 	}
 }
