+++ includes/validation/class-validate-merge-fields.php
@@ -16,6 +16,28 @@ use stdClass;
  * Provides methods for validating merge fields.
  */
 class Validate_Merge_Fields {
+
+	/**
+	 * Create WP_Error objects via factory
+	 *
+	 * Creating WP_Error objects via factory to allow for unit testing
+	 *
+	 * @var Callable
+	 */
+	private $wp_error_factory;
+
+	/**
+	 * Initialize with factory to allow for better unit testing
+	 *
+	 * @param Callable $wp_error_factory Factory to create WP_Error objects
+	 */
+	public function __construct( $wp_error_factory = null ) {
+		// Default to return WP_Error objects
+		$this->wp_error_factory = $wp_error_factory ?? function ( string $code, string $message, $data = null ) {
+			return new WP_Error( $code, $message, $data );
+		};
+	}
+
 	/**
 	 * Validate phone
 	 *
@@ -57,7 +79,7 @@ class Validate_Merge_Fields {
 					esc_html__( '%s must contain the correct amount of digits', 'mailchimp' ),
 					esc_html( $data['name'] )
 				);
-				$opt_val = $this->create_wp_error( 'mc_phone_validation', $message );
+				$opt_val = call_user_func( $this->wp_error_factory, 'mc_phone_validation', $message );
 				break;
 
 			/**
@@ -69,7 +91,7 @@ class Validate_Merge_Fields {
 					esc_html__( '%s must consist of only numbers', 'mailchimp' ),
 					esc_html( $data['name'] )
 				);
-				$opt_val = $this->create_wp_error( 'mc_phone_validation', $message );
+				$opt_val = call_user_func( $this->wp_error_factory, 'mc_phone_validation', $message );
 				break;
 
 			/**
@@ -94,7 +116,7 @@ class Validate_Merge_Fields {
 			if ( empty( $opt_val['addr1'] ) || empty( $opt_val['city'] ) ) {
 				/* translators: %s: field name */
 				$message = sprintf( esc_html__( 'You must fill in %s.', 'mailchimp' ), esc_html( $data['name'] ) );
-				$error   = $this->create_wp_error( 'invalid_address_merge', $message );
+				$error   = call_user_func( $this->wp_error_factory, 'invalid_address_merge', $message );
 				return $error;
 			}
 			/**
@@ -119,31 +141,4 @@ class Validate_Merge_Fields {
 		$merge->country = $opt_val['country'];
 		return $merge;
 	}
-
-	/**
-	 * Creates a WP_Error object for merge field validation errors.
-	 * 
-	 * Enables tests to make assertions regarding error messages
-	 *
-	 * @param string $code    Error code.
-	 * @param string $message Error message.
-	 * @param mixed  $data    Optional. Error data. Default is null.
-	 *
-	 * @return WP_Error
-	 */
-	private function create_wp_error( string $code, string $message ): WP_Error {
-		return new WP_Error( $code, $message );
-	}
-}
-
-// Expose methods as functions to maintain backward compatibility
-// TODO: Introduce a DI Container (PHP DI) to manage dependency injection
-function merge_validate_phone( $opt_val, $data ) {
-	$validator = new Validate_Merge_Fields();
-	return $validator->validate_phone( $opt_val, $data );
-}
-
-function merge_validate_address( $opt_val, $data ) {
-	$validator = new Validate_Merge_Fields();
-	return $validator->validate_address( $opt_val, $data );
 }
